longer and stat merged into master

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,9 +25,9 @@ public class SalesManager {
         return min;
     }
 
-    public int calculateCutOffAverage () {
-        int average = 0;
-        for (int sale : sales) {
+    public long calculateCutOffAverage () {
+        long average = 0;
+        for (long sale : sales) {
             average += sale;
         }
         average = (average - min() - max()) / sales.length;
